BUG: Getting rid of warning of passing size_t to int in SwapVoidRange.

+++ VTK/IO/vtkCosmoReader.cxx
@@ -499,13 +499,13 @@ void vtkCosmoReader::ReadFile(vtkUnstructuredGrid *output)
 #ifdef VTK_WORDS_BIG_ENDIAN
     if(this->ByteOrder == FILE_LITTLE_ENDIAN)
       {
-      vtkByteSwap::SwapVoidRange(block, numFloats, sizeof(float_t));
+      vtkByteSwap::SwapVoidRange(block, numFloats, (int)sizeof(float_t));
       vtkByteSwap::SwapVoidRange(iBlock, numInts, tagBytes);
       }
 #else
     if(this->ByteOrder == FILE_BIG_ENDIAN)
       {
-      vtkByteSwap::SwapVoidRange(block, numFloats, sizeof(float_t));
+      vtkByteSwap::SwapVoidRange(block, numFloats, (int)sizeof(float_t));
       vtkByteSwap::SwapVoidRange(iBlock, numInts, tagBytes);
       }
 #endif
